reorganize imaging_lib into mri module
