+++ pyproject.toml
@@ -81,9 +81,9 @@ max-doc-length = 100
 include = [
     "python/tests",
     "python/lib/db",
-    "python/lib/imaging_lib",
     "python/lib/import_bids_dataset",
     "python/lib/import_dicom_study",
+    "python/lib/mri",
     "python/lib/physio",
     "python/lib/config.py",
     "python/lib/config_file.py",
@@ -91,6 +91,7 @@ include = [
     "python/lib/get_session_info.py",
     "python/lib/logging.py",
     "python/lib/make_env.py",
+    "python/lib/parameter.py",
     "python/scripts/import_dicom_study.py",
     "python/scripts/summarize_dicom_study.py",
     "python/loris_bids_reader",
