+++ python/lib/parameter.py
@@ -1,17 +1,10 @@
-from typing import Literal
-
 from lib.db.models.parameter_type import DbParameterType
 from lib.db.models.parameter_type_category_rel import DbParameterTypeCategoryRel
 from lib.db.queries.parameter_type import get_parameter_type_category_with_name, try_get_parameter_type_with_name_source
 from lib.env import Env
 
 
-def get_or_create_parameter_type(
-    env: Env,
-    parameter_name: str,
-    category: Literal['Electrophysiology Variables', 'MRI Variables'],
-    source: Literal['parameter_file', 'physiological_parameter_file']
-) -> DbParameterType:
+def get_or_create_parameter_type(env: Env, parameter_name: str, category: str, source: str) -> DbParameterType:
     """
     Get a parameter type using its name, or create that parameter if it does not exist.
     """
@@ -24,7 +17,7 @@ def get_or_create_parameter_type(
         name        = parameter_name,
         alias       = None,
         data_type   = 'text',
-        description = f'{parameter_name} created by the lib.imaging.parameter Python module',
+        description = f'{parameter_name} created by the lib.parameter Python module',
         source_from = source,
         queryable   = False,
     )
